Use native_unit_of_measurement to prevent recursion.
In RctPowerSensorEntity the device_class() method used to use unit_of_measurement() to get the unit of measurement for the entity. This no longer works, as with HA 2022.5.0 this implies a recursion back to unit_of_measurement() within the HA device_class() method to figure out if the entity is temperature based.
Thus using native_unit_of_measurement() instead.