Improve battery minimum state of charge.
We actually have two different minimum state of charge settings. One is for
regular use, the other one is to be used if the system is in isolated operation
without grid connectivity. So far the regular one was missing at all.
This change adds the more common missing entity and clarifies the difference.