Some minor improvements (#408)
* Add information about include directory.

Minor fix to make poetry packaging work.

Fixes error:
  poetry build
  [...]
  No file/folder found for package rct-power

* Add entity for external power input.

Additional power source with S0 wired to I/O-board.
This value is required to calculate the actual house total
energy consumption.

* Improve battery minimum state of charge.

We actually have two different minimum state of charge settings. One is for
regular use, the other one is to be used if the system is in isolated operation
without grid connectivity. So far the regular one was missing at all.
This change adds the more common missing entity and clarifies the difference.

* Fix mardown formatting

---------

Co-authored-by: Felix Stürmer <weltenwort@users.noreply.github.com>